3-1Flask/requirements.txt
blinker==1.9.0
click==8.1.7
coloredlogs==15.0.1
Flask==3.1.0
Flask-Cors==4.0.0
flatbuffers==24.3.25
humanfriendly==10.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
packaging==24.2
patchify==0.2.3
pillow==11.0.0
protobuf==5.29.1
sympy==1.13.3
Werkzeug==3.1.3